requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.13.0
pypdf==3.17.1
tenacity==8.2.3
python-dotenv==1.0.0
python-multipart==0.0.6